+++ onchain/src/TrustlessSidechain/MerkleRootTokenMintingPolicy.hs
@@ -20,6 +20,7 @@ import Plutus.V2.Ledger.Api (
   Credential (ScriptCredential),
   CurrencySymbol,
   Datum (getDatum),
+  LedgerBytes (LedgerBytes, getLedgerBytes),
   OutputDatum (OutputDatum),
   Script,
   ScriptContext,
@@ -70,8 +71,9 @@ serialiseMte = Builtins.serialiseData . IsData.toBuiltinData
  > PlutusTx.Builtins.blake2b_256 . PlutusTx.Builtins.serialiseData . PlutusTx.IsData.Class.toBuiltinData
 -}
 {-# INLINEABLE serialiseMrimHash #-}
-serialiseMrimHash :: MerkleRootInsertionMessage -> BuiltinByteString
-serialiseMrimHash = Builtins.blake2b_256 . Builtins.serialiseData . IsData.toBuiltinData
+serialiseMrimHash :: MerkleRootInsertionMessage -> LedgerBytes
+serialiseMrimHash =
+  LedgerBytes . Builtins.blake2b_256 . Builtins.serialiseData . IsData.toBuiltinData
 
 {- | 'mkMintingPolicy' verifies the following
 
@@ -114,7 +116,7 @@ mkMintingPolicy
       ownCurrencySymbol :: CurrencySymbol
       ownCurrencySymbol = Contexts.ownCurrencySymbol ctx
       ownTokenName :: TokenName
-      ownTokenName = Value.TokenName merkleRoot
+      ownTokenName = Value.TokenName $ getLedgerBytes merkleRoot
       sc :: SidechainParams
       sc = get @"sidechainParams" smrm
       committeeDatum :: UpdateCommitteeHashDatum
@@ -152,7 +154,7 @@ mkMintingPolicy
       -- 4., 5. resp. in the documentation of this function.
       p1 = case previousMerkleRoot of
         Nothing -> True
-        Just tn ->
+        Just (LedgerBytes tn) ->
           -- Checks if any of the reference inputs have at least 1 of the last
           -- merkle root.
           let go :: [TxInInfo] -> Bool