+++ onchain/src/TrustlessSidechain/UpdateCommitteeHash.hs
@@ -10,6 +10,7 @@ import Ledger.Value qualified as Value
 import Plutus.Script.Utils.V2.Typed.Scripts qualified as ScriptUtils
 import Plutus.V2.Ledger.Api (
   Datum (getDatum),
+  LedgerBytes (LedgerBytes),
   TokenName (TokenName),
   Value (getValue),
  )
@@ -166,7 +167,7 @@ mkUpdateCommitteeHashValidator uch dat red ctx =
        in verifyMultisig
             (getSidechainPubKey <$> get @"committeePubKeys" red)
             threshold
-            (Builtins.blake2b_256 (serialiseUchm message))
+            (LedgerBytes (Builtins.blake2b_256 (serialiseUchm message)))
             (committeeSignatures red)
     isCurrentCommittee :: Bool
     isCurrentCommittee = aggregateCheck (get @"committeePubKeys" red) $ committeeHash dat
@@ -180,7 +181,7 @@ mkUpdateCommitteeHashValidator uch dat red ctx =
       -- merkle root tokens.
       case get @"previousMerkleRoot" red of
         Nothing -> True
-        Just tn ->
+        Just (LedgerBytes tn) ->
           let go :: [TxInInfo] -> Bool
               go (txInInfo : rest) =
                 ( ( Value.valueOf